% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultar_agenda.R
\name{consultar_agenda}
\alias{consultar_agenda}
\title{Consultar agenda}
\usage{
consultar_agenda(data = NULL, pessoa = NULL)
}
\arguments{
\item{data}{Data que se deseja pesquisar}

\item{pessoa}{Pessoa que se deseja pesquisar}
}
\value{
Retorna uma tibble com 4 colunas: dia do compromisso,
horário do compromisso, descrição do compromisso e local do compromisso
}
\description{
Consulta a agenda de uma pessoa da direção da Anvisa em determinado dia
}
